refactor: model 내부에 step2패키지 및 enum패키지 생성
말씀 하신대로 model 안에 step2 패키지를 만들고, enums도 따로 패키지로 만들었습니다.
