+++ app/src/main/java/com/kau/ttokttok/domain/model/step2/enums/NoiseRange.kt
@@ -1,0 +1,8 @@
+package com.kau.ttokttok.domain.model.step2.enums
+
+// 알림 범위를 구분하는 enum 클래스
+enum class NoiseRange {
+    SURROUNDING,      // 주변 세대 (ex. 상하좌우 8세대)
+    BUILDING          // 동 전체 (ex. 20세대)
+}
+
