+++ app/src/main/java/com/kau/ttokttok/domain/model/step2/enums/ResponseType.kt
@@ -1,0 +1,10 @@
+package com.kau.ttokttok.domain.model.step2.enums
+
+// 이웃의 응답 종류를 나타내는 enum 클래스
+enum class ResponseType {
+    HEARD,    // 네, 저도 들려요
+    QUIET,    // 아니요, 조용해요
+    SORRY,    // 제가 주의할게요
+    PENDING   // 응답 대기 중
+}
+
